Separate layer admin tests to own class
Refactor layer admin tests to use common base class.
